requirements/development.txt
# django
Django==3.1.3
django-simple-history==2.12.0
django-taggit==1.3.0
django-filter==2.4.0
djangorestframework==3.12.2
# django-crispy-forms==1.9.2

# development
django-debug-toolbar==3.1.1
django-extensions==3.0.9

# testing
# pytest-django==4.1.0
# pytest-xdist==2.1.0
# pytest-cov==2.10.1
# Faker==4.14.2

# https://django-bleach.readthedocs.io/en/latest/settings.html#settings
# django-bleach==0.6.1

# database
psycopg2==2.8.6
# pgcli==3.0.0

# utilities
requests==2.24.0
python-dotenv==0.15.0
Markdown==3.3.3
PyYAML==5.3.1
python-dotenv==0.15.0
html2text==2020.1.16
# https://django-bleach.readthedocs.io/en/latest/settings.html#settings
django-bleach==0.6.1
beautifulsoup4==4.9.3

# code
flake8==3.8.4
mypy==0.790

# monitoring
# sentry-sdk==0.19.3
